modint value

+++ source/math_number/modint.cpp
@@ -2,7 +2,7 @@ template<int MOD> struct Modint {
     long long val;
     constexpr Modint(long long v = 0) noexcept : val(v % MOD) { if (val < 0) val += MOD; }
     constexpr int mod() const { return MOD; }
-    constexpr int value() const { return val; }
+    constexpr long long value() const { return val; }
     constexpr Modint operator - () const noexcept { return val ? MOD - val : 0; }
     constexpr Modint operator + (const Modint& r) const noexcept { return Modint(*this) += r; }
     constexpr Modint operator - (const Modint& r) const noexcept { return Modint(*this) -= r; }
